backend/requirements.txt
fastapi>=0.115.7
uvicorn[standard]>=0.34.0
pydantic[email]>=2.5.0
httpx>=0.28.1
python-multipart>=0.0.9
pyjwt>=2.10.1
cryptography>=44.0.0
python-dotenv>=1.0.0
supabase>=2.0.0
openai>=1.0.0
requests>=2.31.0
twilio>=8.0.0
mailchimp-marketing>=3.0.75
